backend/requirements.txt
fastapi>=0.109
uvicorn
pydantic>=2.0
pydantic-settings
langgraph>=0.2
langsmith
openai>=1.0
anthropic>=0.20
redis
celery
httpx
python-dotenv
